REFACTOR ISA-14: back Button View Modifier 적용
