+++ RecordManagment/Presentation/View/OnBoarding/SectionFourView.swift
@@ -41,18 +41,10 @@ struct SectionFourView: View {
             Spacer()
         }
         .navigationBarBackButtonHidden(currentProgress == .goal)
-        .toolbar {
-            ToolbarItem(placement: .topBarLeading) {
-                  Button(action: {
-                      // prev 상태로 이동
-                      withAnimation {
-                          currentProgress = .birth
-                      }
-                  }) {
-                      Image(systemName: "chevron.left")
-                          .higBackSize()
-                          .foregroundStyle(Color.Gray._900())
-                  }
+        .seeDayToolBar {
+            // prev 상태로 이동
+            withAnimation {
+                currentProgress = .birth
             }
         }
     }
