+++ RecordManagment/Presentation/View/OnBoarding/SectionTwoView.swift
@@ -70,18 +70,9 @@ struct SectionTwoView: View {
             Spacer()
         }
         .navigationBarBackButtonHidden(currentProgress == .name)
-        .toolbar {
-            ToolbarItem(placement: .topBarLeading) {
-                  Button(action: {
-                      // prev 상태로 이동
-                      withAnimation {
-                          currentProgress = .record
-                      }
-                  }) {
-                      Image(systemName: "chevron.left")
-                          .higBackSize()
-                          .foregroundStyle(Color.Gray._900())
-                  }
+        .seeDayToolBar {
+            withAnimation {
+                currentProgress = .record
             }
         }
     }
